backend/requirements.txt
# EcoCart Backend Dependencies
# Python 3.11+

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# Database
motor==3.6.0
pymongo==4.9.2
redis==5.2.1

# Data Validation
pydantic==2.10.3
pydantic[email]==2.10.3
pydantic-settings==2.7.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
cryptography==44.0.0
google-auth==2.49.1
requests==2.32.3

# AI Integration
google-generativeai==0.8.3

# Payment Processing

# Image Processing
Pillow==11.0.0
pytesseract==0.3.10

# HTTP Client
httpx==0.28.1

# Environment Management
python-dotenv==1.0.1

# Rate Limiting
slowapi==0.1.9
limits==3.14.0

# Email & Notifications
aiosmtplib==3.0.1

# Cloud Storage
boto3==1.35.93
botocore==1.35.93

# Background Tasks
celery==5.4.0

# Monitoring & Logging
sentry-sdk[fastapi]==2.19.2
structlog==24.4.0

# PDF Generation (for carbon offset certificates)
reportlab==4.2.5

# Web Scraping (optional)
beautifulsoup4==4.12.3

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
httpx==0.28.1
mongomock==4.3.0
fakeredis==2.26.2

# Code Quality
mypy==1.13.0
ruff==0.8.6
